Convert URI templates to regular expressions.
Will match these against the rsync output to find new distributions. The way
it's done with URI::Template and File::Spec is a little circuitous, but I
think correct. A bunch of tests confirm it for the available data.

Changed the rsync output format to omit text indicating hard and symbolic
links. I don't think that matters (and there shouldn't be any, anyway).