Add `update_owner()` method to Indexer.
Updates the owner metadata file with the abstract and date of the latest
release. Will have to go back to PGXN::Manager and add the date to the
metadata file (can't believe I didn't think of it before).

In passing, create the by/ directories when the doc root is created.