Use an rsync log instead of STDOUT.
I ran into issues where a file name was printed to STDOUT and collected by the
app from the pipe before the file was actually copied over or written to its
final destination. This was endlessly annoying. I tried a number of different
approaches to get it to work, based on this bit from the rsync man page:

> Rsync will output the out-format string prior to a file's transfer unless
> one of the transfer-statistic escapes is requested, in which case the
> logging is done at the end of the file's transfer.

It appears that `%b` is such a transfer-statistic escape, but even using it
and `--inplace`, I still got file missing errors. So I gave up. Instead,
execute rsync, then parse its log file instead. This actually simplifies the
code somewhat. It may be less efficient, but it seems to be more reliable.