Add a root-level index.html file.
Copy it from `var/index.html` when a doc root is created. Have the router
serve it when / is requested (but only /).

In passing, only require File::Path in PGXN::API if we're actually using it.
And get rid of the test.html file; just use this index.html for tests,
instead.