Show latest stable version and simplify release merging.
In the /by/dist file, set the version to the latest *stable* version. Fall
back on testing or unstable only if there is no stable distribution.

Also, refactor the merging/updating of user and tag release data. Just rely on
the versions listed on the mirror, rather than what was pulled from the
distribution, since a user won't have released all versions of a distribution
and won't have had the same tags on every distribution. The only thing that
needs to be added to the release data for these two files is the abstract, so
that actually simplifies things greatly, and we do a better job of relying on
what the mirror gives us (as generated by PGXN::Manager).

To properly test this, add fixture files for users and keywords, so that they
are properly udpated from "updates" from the mirror.