Clean doc HTML output with XML::LibXML
Write out the contents of the body only. Use the "provides" field for the
title, if possible. Otherwise use XPath to try to grab the title for first h1
in the HTML document.

Also add code to clean the HTML of all (hopefully!) unwanted markup. This code
is donated from a Lunar/Theory project. Will add tests for it next.