Use URI Template `{+path}` expression.
This will prevent slashes from being escaped, which is essential for doc
paths. I've also switched to "path" instead of "doc" since, well, it's a path.

For now, URI::Template does not support the `{+var}` syntax, so we fake it
for now by:

* Passing the parameter as "+path" instead of "path"
* Locallying preventing URI::Escape from escaping slashes

This is a bit of a hack, but I think that the template is correct, so it's
better to have it now and to modify the code to work with a hypothetical new
version of URI::Template that supports v4 templates later.