Create separate indexes for each doc type.
Rather than one big index. Studying the CPAN search interface a bit, I
realized that its "All" search wasn't actually all of distributions, modules
and authors. First time I tried that with the single big index, I saw that I
was getting duplicate hits (a search found both a distribution and its
included extension). The "All" search is actually a search of "All
documentation". It also indexes the distribution name, so that's part of
the hit, but the results returned are pointers to docs.

So I've done the same thing here. We now have five full text indexes:

* Docs. Indexes full text of documentation, plus title, abstract, and
  distribution name.
* Distributions. Indexes distribution name, abstract, description, and tags.
* Extensions. Indexes name and abstract.
* User. Indexes name, nickname, email, uri, and other details.
* Tag. Index the tag name.

Quite a bit of refactoring in this change, but the nice thing is that all
documentation is now properly indexed.