Test the indexing of distributions.
Running an actual test, found that the code needs to:

* Instantiate new indexers every time, because `commit()` invalidates them.
* Instantiate new searchers for every request, or else miss out on changes
  triggered by any new indexing activity since the last time they were used.

Need to get with the KinoSearch guys to find out if this is a smart or dumb
way to do it. Might be that KS wants to index as much as possible and commit
at the end. It would be annoying that transactions can't be per distribution
without kiling performance, but if that's the way it is, I'll change it. Will
have to decide whether to instantate a new PGXN::API::Searcher for every
request, too, or whether it can be invalidated somehow after an index update.

Meanwhile, I need to come back and test doc, extension, user, and tag indexes.
Will do that next.