Start PGXN::API::Stats.
Will be used to track stats about the various objects and generate summarizing
JSON files with those stats. Plan to use the one for tags for a tag cloud. But
might also have recent extensions listed in the extensions one. Will see.
Development to continue over the next day or two.