Merge user JSON files.
It used to be that user JSON files were only updated when a user released a
distribution. Now they get updated any time a user modifies their data in
PGXN::Manager. User JSON files are also created for new users. So we need to
make sure to always merge in user JSON files, even when the user has no
distributions.

The new `merge_user()` method in Indexer does this. Sync captures the user's
nickname when it spots a user JSON file in the rsync log and passes it off to
this method.

While at it, cache the user's name when the file is read in, so that we have
it in case any distributions are processed for that user after the user file
has been merged. Also, we update the full text index entry for the user.

To make sure this stuff sticks, `finalize()` now calls `_commit()`, so that
any date left to be indexed after the sync complees will be committed to the
index.