Renamed `doc` template and keys.
The template is now `htmldoc`. The key in the extensions JSON and in
`provides` in `META.json` (and template variable name) is now `docpath`. This
is to distinguish it from `docfile`, which is the key used in `provides` for
the file containing the documentation (and which will generally have a file
name suffix like `.txt` or `.md`).

The template is renamed to `htmldoc` in case we ever decide to support other
documentation formats in the API.