Add README to docs index if no other docs found
If a distribution has a README and no other docs recognized by
Text::Markup, add the README to the index. Previously the indexer only
added the README to the distribution docs. It still does, but now also
the docs index, because otherwise a docs search won't turn up results
for a distribution with no other docs.

Add a test to be sure it actually indexes a README when there are no
other docs.

Fixes #12.