Always respect the docfile field
The indexer wasn't really paying attention to the `docfile` field in
`provides` objects, other than to be sure its value was a file to try
to index. Instead is just looked for doc files where the file name was
the same as the extension name. This works okay for distributions that
follow that rule strictly. But many distributions point to the
`README` as the doc file (as mentioned in #12), and others point to
completely different file (as in #10).

So capture much earlier that an extension has pointed to a specific
doc file, and track that relationship throughout the indexing process.
Add test cases to be sure it works.

Resolves #10.