Always set member perms when unzipping
Re-create `t/root/dist/pair/0.1.1/pair-0.1.1.zip` with no permissions
for groups or others to any file or directory and add tests to make sure
they are readable by all and that directories are executable by all.
This causes failures!

Change `unzip` to *always* set file permissions, with different values
for directories (which must be executable) and files (which need not be
executable for serving by the API). Fixes #15.

As a result of the re-organization of `pair-0.1.1.zip`, change the
`find_docs` tests to sort results before comparing them. Apparently
different zip files can index their members in different orders. So just
avoid the problem for the future.