More testing or stable indexing and updates
Be sure to use the testing abstract and version for extension indexing
information when there is no stable version, or the unstable data when
there is no testing. Also prevents passing undefined values to the
Lucy::Index::Indexer.

Similarly, update a user's JSON information with a testing release when
there is no stable release, or the unstable release when no testing.
This way the most recent release is always associated with the user,
rather than no release when there are no existing stable releases!