Parse plain text files as documentation
`README` files were always recognized as documentation for any and all
extension; now the indexer also recognizes and indexes files ending in
`.text` or `.txt`. It registers the Text::Markup::None parser to format
them; their contents will be saved in a `<pre>` block in the resulting
HTML file. Closes #13.